Rename MapGrid and use Layout instead
