+++ cartoframes/assets/style/layout.j2
@@ -1,0 +1,62 @@
+<style>
+.layout-map-cell p {
+  font-family: Roboto;
+  font-size: 14px;
+}
+
+.layout-map.as-main {
+  width: 100%;
+}
+
+.layout-map-cell {
+  align-items: center;
+  display: flex;
+  flex-direction: column;
+  height: 250px;
+  justify-content: center;
+  width: 100%;
+  padding: 8px;
+  position: relative;
+}
+
+.layout-map-cell-legend {
+  flex-direction: row;
+}
+
+h2.as-widget-header__header {
+  font-size: 12px;
+  line-height: 12px;
+  padding: 8px 0;
+}
+
+.layout-map {
+  background-color: silver;
+  height: 75%;
+  width: 85%;
+}
+
+.layout-map > .as-map-area,
+.map-image > .as-map-area {
+  height: 100%;
+}
+
+.layout-map-container {
+  align-items: center;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-around;
+  width: 100%;
+  padding: 0 8px;
+}
+
+.layout-map-row {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-around;
+  width: 100%;
+}
+
+.layout-legends {
+  background-color: white;
+}
+</style>
