--- cartoframes/assets/style/map_grid.html.j2
@@ -1,62 +1,0 @@
-<style>
-.grid-map-cell p {
-  font-family: Roboto;
-  font-size: 14px;
-}
-
-.grid-map.as-main {
-  width: 100%;
-}
-
-.grid-map-cell {
-  align-items: center;
-  display: flex;
-  flex-direction: column;
-  height: 250px;
-  justify-content: center;
-  width: 100%;
-  padding: 8px;
-  position: relative;
-}
-
-.grid-map-cell-legend {
-  flex-direction: row;
-}
-
-h2.as-widget-header__header {
-  font-size: 12px;
-  line-height: 12px;
-  padding: 8px 0;
-}
-
-.grid-map {
-  background-color: silver;
-  height: 75%;
-  width: 85%;
-}
-
-.grid-map > .as-map-area,
-.map-image > .as-map-area {
-  height: 100%;
-}
-
-.grid-map-container {
-  align-items: center;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-around;
-  width: 100%;
-  padding: 0 8px;
-}
-
-.grid-map-row {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-around;
-  width: 100%;
-}
-
-.grid-legends {
-  background-color: white;
-}
-</style>
