--- cartoframes/assets/templates/viz/grid.html.j2
@@ -1,4 +1,0 @@
-{% extends "templates/iframe.html.j2" %}
-{% block srcDoc %}
-  {% include 'templates/viz/main_grid.html.j2' %}
-{% endblock %}
