+++ cartoframes/assets/templates/viz/layout.j2
@@ -1,0 +1,4 @@
+{% extends "templates/iframe.html.j2" %}
+{% block srcDoc %}
+  {% include 'templates/viz/main_layout.html.j2' %}
+{% endblock %}
