+++ cartoframes/viz/__init__.py
@@ -5,7 +5,7 @@ from __future__ import absolute_import
 from .basemaps import Basemaps as basemaps
 from .themes import Themes as themes
 from .map import Map
-from .mapgrid import MapGrid
+from .layout import Layout
 from .layer import Layer
 from .source import Source
 from .style import Style
@@ -19,7 +19,7 @@ __all__ = [
     'basemaps',
     'themes',
     'Map',
-    'MapGrid',
+    'Layout',
     'Layer',
     'Source',
     'Style',
