+++ cartoframes/viz/constants.py
@@ -94,4 +94,4 @@ CLUSTER_OPERATIONS = {
 
 THEMES = ['dark', 'light']
 
-DEFAULT_GRID_M_SIZE = 1
+DEFAULT_LAYOUT_M_SIZE = 1
